Update Action Table

+++ src/main/java/Swing/Table/EventAction_Impl.java
@@ -5,24 +5,33 @@
 package Swing.Table;
 
 import AdminForm.Drug.DetailsDrug;
+import DatabaseAccessObject_DAO.Drug_Dao;
+import DatabaseAccessObject_DAO.MedicalExamination_Dao;
 import DatabaseAccessObject_DAO.Patient_Dao;
+import DatabaseAccessObject_DAO.Service_Dao;
 import DatabaseAccessObject_DAO.Staff_Dao;
+import DatabaseAccessObject_Impl.Drug_DaoImpl;
+import DatabaseAccessObject_Impl.MedicalExamination_DaoImpl;
 import DatabaseAccessObject_Impl.Patient_DaoImpl;
+import DatabaseAccessObject_Impl.Service_DaoImpl;
 import DatabaseAccessObject_Impl.Staff_DaoImpl;
 import Dialog.Swal_Confirm;
 import Dialog.Swal_Notification;
 import Enum.RoleName;
 import Enum.TypeInterface;
 import Enum.TypeNotification;
+import Form.Drug.NewDrug;
 import Form.Employee.EmployeeProfile;
 import Form.Employee.NewEmployee;
 import Form.Medical.NewMedical;
 import Form.Patient.NewPatient;
 import Form.Patient.PatientDetails;
+import Form.Service.NewService;
 import Model.Drug;
 import Model.Employee;
 import Model.MedicalExamination;
 import Model.Patient;
+import Model.Service;
 import ViewForm.Main;
 import java.awt.Component;
 import javax.swing.JPanel;
@@ -50,29 +59,7 @@ public abstract class EventAction_Impl implements EventAction{
             Patient patient = (Patient)obj;
             System.out.println(".delete()" + patient.getID());
             if (row != -1) {
-                if(!this.employeecurrent.getRoleName().equals(RoleName.Admin.toString()))
-                    showMessageError("<html><div style='text-align:center;'>You do not have permission to delete <br>Patient : " 
-                            + patient.getFullName() + "(ID: " + patient.getID() + ")</div></html>", TypeNotification.Error);
-                else 
-                {
-                    showMessageError("Can't Delete Patient", TypeNotification.Default);
-//                    if(showMessage("Delete Patient : " + patient.getFullName() + "(ID: " + patient.getID() + ")", TypeNotification.Default)) {
-//                        // Xóa sinh viên tại vị trí rowIndex trong dữ liệu của JTable
-//                        int PatientID = patient.getID();
-//                        //Hàm Xóa
-//                        System.out.println("UserId = " + PatientID);
-//
-//                        if (this.table.isEditing()) {
-//                            this.table.getCellEditor().stopCellEditing();
-//                        }
-//                        DefaultTableModel model = (DefaultTableModel) table.getModel();
-//                        model.removeRow(row);
-//                        
-//                    
-//                    } else {
-//                        System.out.println("User click Cancel");
-//                    }
-                }
+                showMessageError("Can't Delete Patient", TypeNotification.Warning);
             } else {
                 System.out.println("No row selected");
             }
@@ -81,21 +68,84 @@ public abstract class EventAction_Impl implements EventAction{
             Employee employee = (Employee) obj;
             System.out.println(".delete()" + employee.getID());
             if (row != -1) {
-                if (showMessage("Delete Employee : " + employee.getFullName() + "(ID: " + employee.getID() + ")", TypeNotification.Default)) {
-                    if (this.table.isEditing()) {
-                        this.table.getCellEditor().stopCellEditing();
+                if(this.employeecurrent.getRoleName().equals(RoleName.Admin.toString()))
+                {
+                    if (showMessage("Delete Employee : " + employee.getFullName() + "(ID: " + employee.getID() + ")", TypeNotification.Default)) {
+                        if (this.table.isEditing()) {
+                            this.table.getCellEditor().stopCellEditing();
+                        }
+                        DefaultTableModel model = (DefaultTableModel) table.getModel();
+                        model.removeRow(row);
+
+                        Staff_Dao employee_Dao = new Staff_DaoImpl();
+                        employee_Dao.Delete_Staff(employee.getID());
                     }
-                    DefaultTableModel model = (DefaultTableModel) table.getModel();
-                    model.removeRow(row);
-                    
-                    Staff_Dao employee_Dao = new Staff_DaoImpl();
-                    employee_Dao.Delete_Staff(employee.getID());
-                } else {
-                    System.out.println("User click Cancel");
+                }
+                else {
+                    showMessageError("You don't have permission to use this function.", TypeNotification.Warning);
                 }
             } else {
                 System.out.println("No row selected");
             }
+        }
+        if (obj instanceof MedicalExamination) {
+            MedicalExamination medical = (MedicalExamination) obj;
+            System.out.println(".delete()" + medical.getID());
+            if (row != -1) {
+                showMessageError("Can't Delete MedicalExamination", TypeNotification.Warning);
+            }
+        }
+        if (obj instanceof Drug) {
+            Drug object = (Drug) obj;
+            System.out.println(".delete()" + object.getID());
+            if (row != -1) {
+                if(this.employeecurrent.getRoleName().equals(RoleName.Admin.toString()) || this.employeecurrent.getRoleName().equals(RoleName.Pharmacist.toString()))
+                {
+                    if (showMessage("Delete Employee : " + object.getDrugName()+ "(ID: " + object.getID() + ")", TypeNotification.Default)) {
+                        if (this.table.isEditing()) {
+                            this.table.getCellEditor().stopCellEditing();
+                        }
+                        DefaultTableModel model = (DefaultTableModel) table.getModel();
+                        model.removeRow(row);
+
+                        Drug_Dao object_Dao = new Drug_DaoImpl();
+                        object_Dao.Delete_Drug(object.getID());
+                    } else {
+                        System.out.println("User click Cancel");
+                    }
+                }
+                else {
+                    showMessageError("You don't have permission to use this function.", TypeNotification.Warning);
+                }
+            } else {
+                System.out.println("No row selected");
+            } 
+        }
+        if (obj instanceof Service) {
+            Service object = (Service) obj;
+            System.out.println(".delete()" + object.getID());
+            if (row != -1) {
+                if(this.employeecurrent.getRoleName().equals(RoleName.Admin.toString()))
+                {
+                    if (showMessage("Delete Service : " + object.getServiceName()+ "(ID: " + object.getID() + ")", TypeNotification.Default)) {
+                        if (this.table.isEditing()) {
+                            this.table.getCellEditor().stopCellEditing();
+                        }
+                        DefaultTableModel model = (DefaultTableModel) table.getModel();
+                        model.removeRow(row);
+
+                        Service_Dao object_Dao = new Service_DaoImpl();
+                        object_Dao.Delete_Service(object.getID());
+                    } else {
+                        System.out.println("User click Cancel");
+                    }
+                }
+                else {
+                    showMessageError("You don't have permission to use this function.", TypeNotification.Warning);
+                }
+            } else {
+                System.out.println("No row selected");
+            } 
         }
     }
     
@@ -106,13 +156,14 @@ public abstract class EventAction_Impl implements EventAction{
             int rowIndex = table.getSelectedRow();
             Patient patient = (Patient)obj;
             if (rowIndex >= 0) {
+                
                 if (showMessage("You are going to the patient edit interface : " + patient.getFullName()+ "(ID: " + patient.getID()+ ")", TypeNotification.Default)) {
-                    // Chuyển giao diện sang Edit
-                    showForm(new NewPatient(main, TypeInterface.Edit, patient, this.employeecurrent));
-                    // ...
+                // Chuyển giao diện sang Edit
+                showForm(new NewPatient(main, TypeInterface.Edit, patient, this.employeecurrent));
                 } else {
                     System.out.println("User click Cancel");
                 }
+                
             } else {
                 System.out.println("No row selected");
             }
@@ -122,12 +173,19 @@ public abstract class EventAction_Impl implements EventAction{
             int rowIndex = table.getSelectedRow();
             Employee object = (Employee)obj;
             if (rowIndex >= 0) {
-                if (showMessage("You are going to the employee edit interface : " + object.getFullName()+ "(ID: " + object.getID()+ ")", TypeNotification.Default)) {
-                    // Chuyển giao diện sang Edit
-                    showForm(new NewEmployee(main, TypeInterface.Edit, object, this.employeecurrent));
-                    // ...
-                } else {
-                    System.out.println("User click Cancel");
+                if(this.employeecurrent.getRoleName().equals(RoleName.Admin.toString()))
+                {
+                    if (showMessage("You are going to the employee edit interface : " + object.getFullName()+ "(ID: " + object.getID()+ ")", TypeNotification.Default)) {
+                        // Chuyển giao diện sang Edit
+                        //Staff_Dao employee_Dao = new Staff_DaoImpl();
+                        showForm(new NewEmployee(main, TypeInterface.Edit, object, this.employeecurrent));
+                        // ...
+                    } else {
+                        System.out.println("User click Cancel");
+                    }
+                }
+                else {
+                    showMessageError("You don't have permission to use this function.", TypeNotification.Warning);
                 }
             } else {
                 System.out.println("No row selected");
@@ -140,11 +198,56 @@ public abstract class EventAction_Impl implements EventAction{
             if (rowIndex >= 0) {
                 if (showMessage("You are going to the medical edit interface : " + object.getPatient().getFullName()+ "(ID: " + object.getID()+ ")", TypeNotification.Default)) {
                     // Chuyển giao diện sang Edit
-                    //Patient_Dao patient_Dao = new Patient_DaoImpl();
+                    //MedicalExamination_Dao medical = new MedicalExamination_DaoImpl();
                     showForm(new NewMedical(main, TypeInterface.Edit, this.employeecurrent, object));
-                    // ...
                 } else {
                     System.out.println("User click Cancel");
+                }
+            } else {
+                System.out.println("No row selected");
+            }
+        }
+        if(obj instanceof Drug)
+        {
+            int rowIndex = table.getSelectedRow();
+            Drug object = (Drug)obj;
+            if (rowIndex >= 0) {
+                if(this.employeecurrent.getRoleName().equals(RoleName.Admin.toString()) || this.employeecurrent.getRoleName().equals(RoleName.Pharmacist.toString()))
+                {
+                    if (showMessage("You are going to the drug edit interface : " + object.getDrugName()+ "(ID: " + object.getID()+ ")", TypeNotification.Default)) {
+                        // Chuyển giao diện sang Edit
+                        //Drug_Dao object_Dao = new Drug_DaoImpl();
+                        showForm(new NewDrug(main, TypeInterface.Edit, this.employeecurrent, object));
+                        // ...
+                    } else {
+                        System.out.println("User click Cancel");
+                    }
+                }
+                else {
+                    showMessageError("You don't have permission to use this function.", TypeNotification.Warning);
+                }
+            } else {
+                System.out.println("No row selected");
+            }
+        }
+        if(obj instanceof Service)
+        {
+            int rowIndex = table.getSelectedRow();
+            Service object = (Service)obj;
+            if (rowIndex >= 0) {
+                if(this.employeecurrent.getRoleName().equals(RoleName.Admin.toString()) || this.employeecurrent.getRoleName().equals(RoleName.Pharmacist.toString()))
+                {
+                    if (showMessage("You are going to the drug edit interface : " + object.getServiceName()+ "(ID: " + object.getID()+ ")", TypeNotification.Default)) {
+                        // Chuyển giao diện sang Edit
+                        //Drug_Dao object_Dao = new Drug_DaoImpl();
+                        showForm(new NewService(main, TypeInterface.Edit, object, this.employeecurrent));
+                        // ...
+                    } else {
+                        System.out.println("User click Cancel");
+                    }
+                }
+                else {
+                    showMessageError("You don't have permission to use this function.", TypeNotification.Warning);
                 }
             } else {
                 System.out.println("No row selected");
@@ -162,8 +265,6 @@ public abstract class EventAction_Impl implements EventAction{
                 if (showMessage("View Patient : " + object.getFullName()+ "(ID: " + object.getID()+ ")", TypeNotification.Default)) {
                     // Chuyển giao diện xem tại đây
                     showForm(new PatientDetails(main, this.employeecurrent, object));
-                    // ...
-                    System.out.println("User click OK");
                 } else {
                     System.out.println("User click Cancel");
                 }
@@ -179,8 +280,6 @@ public abstract class EventAction_Impl implements EventAction{
                 if (showMessage("View Patient : " + object.getFullName()+ "(ID: " + object.getID()+ ")", TypeNotification.Default)) {
                     // Chuyển giao diện xem tại đây
                     showForm(new EmployeeProfile(main, this.employeecurrent, object));
-                    // ...
-                    System.out.println("User click OK");
                 } else {
                     System.out.println("User click Cancel");
                 }
@@ -191,16 +290,17 @@ public abstract class EventAction_Impl implements EventAction{
         else if(obj instanceof Drug)
         {
             int rowIndex = table.getSelectedRow();
-            Drug object = (Drug)obj;
             if (rowIndex >= 0) {
-                if (showMessage("View Patient : " + object.getDrugName()+ "(ID: " + object.getID()+ ")", TypeNotification.Default)) {
-                    // Chuyển giao diện xem tại đây
-                    //showForm(new DetailsDrug(main, this.employeecurrent, object));
-                    // ...
-                    System.out.println("User click OK");
-                } else {
-                    System.out.println("User click Cancel");
-                }
+                showMessageError("The fuction doesn't exits, it look like it's under construction.", TypeNotification.Warning);
+            } else {
+                System.out.println("No row selected");
+            }
+        }
+        else if(obj instanceof Service)
+        {
+            int rowIndex = table.getSelectedRow();
+            if (rowIndex >= 0) {
+                showMessageError("The fuction doesn't exits, it look like it's under construction.", TypeNotification.Warning);
             } else {
                 System.out.println("No row selected");
             }
